pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lego-sorter"
version = "1.0.0"
description = "Blender-based simulation for sorting LEGO parts"
authors = [
    {name = "LEGO Sorter Project"}
]
requires-python = ">=3.10"
dependencies = [
    "fake-bpy-module-latest==20250630"
]

[tool.setuptools]
# Prevent setuptools from auto-discovering packages in this flat repo layout.
packages = []

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
    "mypy",
    "ruff"
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "bpy.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "mathutils.*"
ignore_missing_imports = true
